ADM-983 [frontend]: refactor
