+++ frontend/src/components/Common/ReportGrid/ReportCard/DoraMetricsDialog/style.tsx
@@ -2,7 +2,6 @@ import { styled } from '@mui/material/styles';
 
 export const DefinitionTitle = styled('div')({
   fontWeight: 'bolder',
-  fontSize: '1rem',
   lineHeight: '2rem',
 });
 
@@ -10,7 +9,6 @@ export const StyledDialogUl = styled('ul')({
   lineHeight: '1.25rem',
   '& span': {
     fontWeight: 'bolder',
-    fontSize: '1rem',
   },
 });
 
@@ -18,8 +16,8 @@ export const StyledDialogLi = styled('li')({
   margin: '2rem 0 0 1rem',
   '& ul': {
     margin: '0 1rem',
-  },
-  '& ul li': {
-    lineHeight: '1.5rem',
+    '& li': {
+      lineHeight: '1.5rem',
+    },
   },
 });
