+++ frontend/src/containers/MetricsStep/Advance/style.tsx
@@ -37,3 +37,14 @@ export const AdvancedWrapper = styled('div')`
   flex-direction: column;
   gap: 1rem;
 `;
+
+export const TitleAndTooltipContainer = styled('div')({
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'center',
+  gap: '0.3125rem',
+});
+
+export const TooltipContainer = styled('div')({
+  marginLeft: '0.25rem',
+});
