+++ frontend/src/containers/MetricsStep/CycleTime/style.tsx
@@ -17,14 +17,3 @@ export const ItemCheckbox = styled(Checkbox)({
   padding: 0,
   marginRight: '0.5rem',
 });
-
-export const TitleAndTooltipContainer = styled('div')({
-  display: 'flex',
-  flexDirection: 'row',
-  alignItems: 'center',
-  gap: '0.3125rem',
-});
-
-export const TooltipContainer = styled('div')({
-  marginLeft: '0.25rem',
-});
